0430 3

+++ src/main/java/controller/CategoryController.java
@@ -1,7 +1,9 @@
 package controller;
 
 import java.util.List;
+import java.util.Map;
 
+import javax.servlet.ServletContext;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import dto.CategoryDto;
 import dto.User;
+import mapper.CategoryMapper;
 import service.CategoryService;
 
 @Controller
@@ -23,7 +26,19 @@ public class CategoryController {
 
     @Autowired
     private CategoryService categoryService;
+    //0430 hto
+    
+    @Autowired
+    private ServletContext servletContext;
+    @Autowired CategoryMapper categoryMapper;
 
+    
+    private void refreshCategoryContext() {
+    	List<Map<String, Object>> updatedList = categoryMapper.selectParentOnly();
+        servletContext.setAttribute("clsList", updatedList);
+        System.out.println("[System] 카테고리 수정으로 인해 메뉴 리스트가 갱신되었습니다.");
+    }
+    
     @GetMapping("/sub-list")
     @ResponseBody
     public List<CategoryDto> getSubList(@RequestParam("parent_code") int parentCode) {
@@ -40,6 +55,7 @@ public class CategoryController {
         }
         
         categoryService.insertCategory(dto);
+        refreshCategoryContext();
         return "success";
     }
 
@@ -47,11 +63,15 @@ public class CategoryController {
     @PostMapping("/admin/delete")
     @ResponseBody
     public String deleteCategory(@RequestParam int category_code, HttpSession session) {
-        // 권한 체크 후 삭제 로직
+        // 권한 체크 후 삭제
+        User loginUser = (User) session.getAttribute("loginUser");
+        if (loginUser == null || loginUser.getUser_role() != 2) return "fail";
         categoryService.deleteCategory(category_code);
+        refreshCategoryContext();
         return "success";
     }
-    @GetMapping("/admin/list") // 또는 원하는 주소
+    
+    @GetMapping("/admin/list") //
     public String categoryMain(Model model, HttpSession session) {
         // 권한 체크 (관리자만 진입 가능)
         User loginUser = (User) session.getAttribute("loginUser");
@@ -63,7 +83,6 @@ public class CategoryController {
         List<CategoryDto> list = categoryService.getAllCategories();
         model.addAttribute("categoryList", list);
 
-        // JSP 파일의 경로 (view/admin/class/category.jsp 라면)
         return "admin/class/category"; 
     }
     
@@ -78,6 +97,7 @@ public class CategoryController {
         }
 
         categoryService.updateCategory(dto); // 업데이트 서비스 호출
+        refreshCategoryContext();
         return "success";
     }
 }
